+++ projects/igniteui-angular-i18n/src/i18n/SV/grid-resources.ts
@@ -109,6 +109,8 @@ const GridResourceStringsSV_: ExpandRequire<IGridResourceStrings> = {
     igx_grid_hiding_uncheck_all_label: 'Dölj alla',
     igx_grid_pinning_check_all_label: 'Fäst alla',
     igx_grid_pinning_uncheck_all_label: 'Lossa alla',
+    igx_grid_row_edit_btn_done: 'Klar',
+    igx_grid_row_edit_btn_cancel: 'Avbryt',
     igx_grid_toolbar_actions_filter_prompt: 'Filtrera kolumnlista ...',
     igx_grid_toolbar_pinning_button_tooltip: 'Öppna rullgardinsmenyn med kolumnfästning',
     igx_grid_toolbar_hiding_button_tooltip: 'Öppna rullgardinsmenyn som döljer kolumnen',